+++ lib/t/t_printpf.m
@@ -35,6 +35,10 @@ if have_feature('octave')
     end
     s1 = warning('query', file_in_path_warn_id);
     warning('off', file_in_path_warn_id);
+else
+    sing_matrix_warn_id = 'MATLAB:nearlySingularMatrix';
+    s2 = warning('query', sing_matrix_warn_id);
+    warning('off', sing_matrix_warn_id);
 end
 
 t = 'printpf : ';
@@ -113,6 +117,8 @@ delete(tmpfnamedc);
 
 if have_feature('octave')
     warning(s1.state, file_in_path_warn_id);
+else
+    warning(s2.state, sing_matrix_warn_id);
 end
 
 t_end;